requirements.txt
django==2.2.5
djangorestframework
django-pyodbc-azure
python-docx
model_mommy